move items up and down MVP
MVP: It does move the items up and down in the internal
representation (the list).
However, this does not necessarily imply that the item moves
on the screen: if the adajcent list item is not a neighbour
but only visible in another part (inCart different), then
there is no immediate effect. Clicking more often will work,
though.